@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk.WindowsDesktop">
   <PropertyGroup>
     <OutputType>WinExe</OutputType>
-    <TargetFrameworks>net452;netcoreapp3.0</TargetFrameworks>
+    <TargetFrameworks>net452;netcoreapp3.1</TargetFrameworks>
     <UseWPF>true</UseWPF>
     <ApplicationIcon>SciChartDemoIcon32.ico</ApplicationIcon>
     <StartupObject />
@@ -15,10 +15,10 @@
     <OutputPath>..\..\Build\Release\</OutputPath>
     <PlatformTarget>AnyCPU</PlatformTarget>
   </PropertyGroup>
-  <PropertyGroup Condition="'$(Configuration)|$(TargetFramework)|$(Platform)'=='Debug|netcoreapp3.0|AnyCPU'">
+  <PropertyGroup Condition="'$(Configuration)|$(TargetFramework)|$(Platform)'=='Debug|netcoreapp3.1|AnyCPU'">
     <OutputPath>..\..\Build\Debug</OutputPath>
   </PropertyGroup>
-  <PropertyGroup Condition="'$(Configuration)|$(TargetFramework)|$(Platform)'=='Release|netcoreapp3.0|AnyCPU'">
+  <PropertyGroup Condition="'$(Configuration)|$(TargetFramework)|$(Platform)'=='Release|netcoreapp3.1|AnyCPU'">
     <OutputPath>..\..\Build\Release</OutputPath>
   </PropertyGroup>
   <ItemGroup>
@@ -80,8 +80,8 @@
     <EmbeddedResource Include="Resources\stopwords.dat" />
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="DotNetProjects.WpfToolkit.Input" Version="6.0.73" />
-    <PackageReference Include="RestSharp" Version="106.6.10" />
+    <PackageReference Include="DotNetProjects.WpfToolkit.Input" Version="6.0.85" />
+    <PackageReference Include="RestSharp" Version="106.11.7" />
     <PackageReference Include="SciChart.Wpf.UI" Version="3.1.251" />
   </ItemGroup>
   <ItemGroup>
@@ -110,15 +110,15 @@
     </None>
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="SciChart" Version="6.2.2.13419-beta" />
+    <PackageReference Include="SciChart" Version="6.2.2.13445-beta" />
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="SciChart3D" Version="6.2.2.13419-beta" />
+    <PackageReference Include="SciChart3D" Version="6.2.2.13445-beta" />
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="SciChart.DirectX" Version="6.2.2.13419-beta" />
+    <PackageReference Include="SciChart.DirectX" Version="6.2.2.13445-beta" />
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="SciChart.DrawingTools" Version="6.2.2.13419-beta" />
+    <PackageReference Include="SciChart.DrawingTools" Version="6.2.2.13445-beta" />
   </ItemGroup>
 </Project>
