@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk.WindowsDesktop">
   <PropertyGroup>
     <OutputType>Library</OutputType>
-    <TargetFrameworks>net452;netcoreapp3.0</TargetFrameworks>
+    <TargetFrameworks>net452;netcoreapp3.1</TargetFrameworks>
     <UseWPF>true</UseWPF>
     <ApplicationIcon />
     <StartupObject />
@@ -15,13 +15,13 @@
     <OutputPath>..\..\Build\Release\</OutputPath>
     <DocumentationFile>..\..\Build\Release\net452\SciChart.Examples.ExternalDependencies.xml</DocumentationFile>
   </PropertyGroup>
-  <PropertyGroup Condition="'$(Configuration)|$(TargetFramework)|$(Platform)'=='Debug|netcoreapp3.0|AnyCPU'">
+  <PropertyGroup Condition="'$(Configuration)|$(TargetFramework)|$(Platform)'=='Debug|netcoreapp3.1|AnyCPU'">
     <OutputPath>..\..\Build\Debug</OutputPath>
-    <DocumentationFile>..\..\Build\Debug\netcoreapp3.0\SciChart.Examples.ExternalDependencies.xml</DocumentationFile>
+    <DocumentationFile>..\..\Build\Debug\netcoreapp3.1\SciChart.Examples.ExternalDependencies.xml</DocumentationFile>
   </PropertyGroup>
-  <PropertyGroup Condition="'$(Configuration)|$(TargetFramework)|$(Platform)'=='Release|netcoreapp3.0|AnyCPU'">
+  <PropertyGroup Condition="'$(Configuration)|$(TargetFramework)|$(Platform)'=='Release|netcoreapp3.1|AnyCPU'">
     <OutputPath>..\..\Build\Release</OutputPath>
-    <DocumentationFile>..\..\Build\Release\netcoreapp3.0\SciChart.Examples.ExternalDependencies.xml</DocumentationFile>
+    <DocumentationFile>..\..\Build\Release\netcoreapp3.1\SciChart.Examples.ExternalDependencies.xml</DocumentationFile>
   </PropertyGroup>
   <ItemGroup>
     <Compile Remove="Controls\SciChartInteractionToolbar\MyCustomAnnotation.cs" />
@@ -67,7 +67,7 @@
     <EmbeddedResource Include="Resources\Data\Waveform.txt.gz" />
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="Microsoft.Xaml.Behaviors.Wpf" Version="1.0.30" />
+    <PackageReference Include="Microsoft.Xaml.Behaviors.Wpf" Version="1.1.19" />
   </ItemGroup>
   <ItemGroup>
   </ItemGroup>
@@ -92,15 +92,15 @@
     </Page>
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="SciChart" Version="6.2.2.13419-beta" />
+    <PackageReference Include="SciChart" Version="6.2.2.13445-beta" />
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="SciChart3D" Version="6.2.2.13419-beta" />
+    <PackageReference Include="SciChart3D" Version="6.2.2.13445-beta" />
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="SciChart.DirectX" Version="6.2.2.13419-beta" />
+    <PackageReference Include="SciChart.DirectX" Version="6.2.2.13445-beta" />
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="SciChart.DrawingTools" Version="6.2.2.13419-beta" />
+    <PackageReference Include="SciChart.DrawingTools" Version="6.2.2.13445-beta" />
   </ItemGroup>
 </Project>
