@@ -2,15 +2,36 @@
 
 namespace SciChart.Examples.ExternalDependencies.Common
 {
+    /// <summary>
+    /// Converts an array of 2 elements to a boolean value which indicates whether they are equal.
+    /// </summary>
     public class IsSameObjectsMultiConverter : IMultiValueConverter
     {
-
+        /// <summary>
+        /// Modifies the source data before passing it to the target for display in the UI.
+        /// </summary>
+        /// <param name="values">The source data array being passed to the target.</param>
+        /// <param name="targetType">The <see cref="T:System.Type"/> of data expected by the target dependency property.</param>
+        /// <param name="parameter">An optional parameter to be used in the converter logic.</param>
+        /// <param name="culture">The culture of the conversion.</param>
+        /// <returns>
+        /// The value to be passed to the target dependency property.
+        /// </returns>
         public object Convert(object[] values, System.Type targetType, object parameter, System.Globalization.CultureInfo culture)
-         {
-            return values.Length == 2 
-                && values[0].Equals(values[1]);
+        {
+            return values?.Length == 2 && values[0] == values[1];
         }
 
+        /// <summary>
+        /// Modifies the target data before passing it to the source object.
+        /// </summary>
+        /// <param name="value">The target data being passed to the source.</param>
+        /// <param name="targetTypes">The <see cref="T:System.Type"/> array of data expected by the source object.</param>
+        /// <param name="parameter">An optional parameter to be used in the converter logic.</param>
+        /// <param name="culture">The culture of the conversion.</param>
+        /// <returns>
+        /// The value to be passed to the source object.
+        /// </returns>
         public object[] ConvertBack(object value, System.Type[] targetTypes, object parameter, System.Globalization.CultureInfo culture)
         {
             return null;
