@@ -16,7 +16,6 @@
 using System;
 using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Threading;
 using SciChart.Charting.Model.DataSeries;
 using SciChart.Charting.Model.DataSeries.Heatmap2DArrayDataSeries;
 using SciChart.Charting.Visuals.RenderableSeries;
@@ -33,34 +32,34 @@ namespace SciChart.Examples.Examples.HeatmapChartTypes.RealTimeHeatmap
         public ContoursExampleView()
         {
             InitializeComponent();
-
-            var colormap = heatmapSeries.ColorMap;
-            var cpMin = colormap.Minimum;
-            var cpMax = colormap.Maximum;
-
-            heatmapSeries.DataSeries = CreateSeries(3, 300, 200, cpMin, cpMax);
-
+            
+            heatmapSeries.DataSeries = CreateSeries(3, 300, 200, heatmapSeries.ColorMap.Maximum);
             contourSeries.DataSeries = heatmapSeries.DataSeries;
-
         }
 
-        private IDataSeries CreateSeries(int index, int width, int height, double cpMin, double cpMax)
+        private IDataSeries CreateSeries(int index, int width, int height, double cpMax)
         {
-            double angle = Math.PI * 2 * index / 30;
-            int w = width, h = height;
+            var w = width;
+            var h = height;
+
+            var angle = Math.PI * 2 * index / 30;
             var data = new double[h, w];
-            double smallValue = 0;
+            var smallValue = 0d;
+            
             for (int x = 0; x < w; x++)
             {
                 for (int y = 0; y < h; y++)
                 {
                     var v = (1 + Math.Sin(x * 0.04 + angle)) * 50 +
                             (1 + Math.Sin(y * 0.1 + angle)) * 50 * (1 + Math.Sin(angle * 2));
+                    
                     var cx = w / 2;
                     var cy = h / 2;
+                    
                     var r = Math.Sqrt((x - cx) * (x - cx) + (y - cy) * (y - cy));
                     var exp = Math.Max(0, 1 - r * 0.008);
                     var zValue = (v * exp);
+                    
                     data[y, x] = (zValue > cpMax) ? cpMax : zValue;
                     data[y, x] += smallValue;
                 }
@@ -73,21 +72,20 @@ namespace SciChart.Examples.Examples.HeatmapChartTypes.RealTimeHeatmap
 
         private void OnApplyMajorStyle(object sender, RoutedEventArgs e)
         {
-            contourSeries.MajorLineStyle = ApplyMajorStyleCkb.IsChecked.Value ? (Style)Resources["MajorContourLineStyle"] : null;
+            contourSeries.MajorLineStyle = ApplyMajorStyleCkb.IsChecked == true ? (Style)Resources["MajorContourLineStyle"] : null;
             sciChart.ZoomExtents();
         }
 
         private void OnApplyMinorStyle(object sender, RoutedEventArgs e)
         {
-            contourSeries.MinorLineStyle = ApplyMinorStyleCkb.IsChecked.Value ? (Style)Resources["MinorContourLineStyle"] : null;
+            contourSeries.MinorLineStyle = ApplyMinorStyleCkb.IsChecked == true ? (Style)Resources["MinorContourLineStyle"] : null;
             sciChart.ZoomExtents();
         }
 
         private void OnApplyPalette(object sender, RoutedEventArgs e)
         {
-            contourSeries.ColorMap = ApplyPaletteCkb.IsChecked.Value ? (HeatmapColorPalette)Resources["ColorPalette"] : null;
+            contourSeries.ColorMap = ApplyPaletteCkb.IsChecked == true ? (HeatmapColorPalette)Resources["ColorPalette"] : null;
             sciChart.ZoomExtents();
         }
-        
     }
 }
