@@ -16,6 +16,7 @@
 using System.Windows;
 using SciChart.Charting.Visuals;
 using SciChart.Charting.Visuals.Axes;
+using SciChart.Core.Extensions;
 
 namespace SciChart.Examples.Examples.ModifyAxisBehaviour
 {
@@ -25,28 +26,28 @@ namespace SciChart.Examples.Examples.ModifyAxisBehaviour
     public class SwitchAxisTypeBehavior
     {
         // Define the Default XAxis -- this can be done in XAML
-        public static readonly DependencyProperty DefaultXAxisProperty = DependencyProperty.RegisterAttached(
-            "DefaultXAxis", typeof(AxisBase), typeof(SwitchAxisTypeBehavior), new PropertyMetadata(default(AxisBase), OnXPropertyChanged));
+        public static readonly DependencyProperty DefaultXAxisProperty = DependencyProperty.RegisterAttached
+            ("DefaultXAxis", typeof(AxisBase), typeof(SwitchAxisTypeBehavior), new PropertyMetadata(default(AxisBase), OnXPropertyChanged));
 
         // Define the Alternative XAxis -- this can be done in XAML
-        public static readonly DependencyProperty AlternativeXAxisProperty = DependencyProperty.RegisterAttached(
-            "AlternativeXAxis", typeof(AxisBase), typeof(SwitchAxisTypeBehavior), new PropertyMetadata(default(AxisBase), OnXPropertyChanged));
+        public static readonly DependencyProperty AlternativeXAxisProperty = DependencyProperty.RegisterAttached
+            ("AlternativeXAxis", typeof(AxisBase), typeof(SwitchAxisTypeBehavior), new PropertyMetadata(default(AxisBase), OnXPropertyChanged));
 
         // Define a boolean attached property which switches between default and alternative XAxis
-        public static readonly DependencyProperty UseAlternateXAxisProperty = DependencyProperty.RegisterAttached(
-            "UseAlternateXAxis", typeof(bool), typeof(SwitchAxisTypeBehavior), new PropertyMetadata(default(bool), OnXPropertyChanged));
+        public static readonly DependencyProperty UseAlternateXAxisProperty = DependencyProperty.RegisterAttached
+            ("UseAlternateXAxis", typeof(bool), typeof(SwitchAxisTypeBehavior), new PropertyMetadata(default(bool), OnXPropertyChanged));
 
         // Define the Default YAxis -- this can be done in XAML
-        public static readonly DependencyProperty DefaultYAxisProperty = DependencyProperty.RegisterAttached(
-            "DefaultYAxis", typeof(AxisBase), typeof(SwitchAxisTypeBehavior), new PropertyMetadata(default(AxisBase), OnYPropertyChanged));
-        
+        public static readonly DependencyProperty DefaultYAxisProperty = DependencyProperty.RegisterAttached
+            ("DefaultYAxis", typeof(AxisBase), typeof(SwitchAxisTypeBehavior), new PropertyMetadata(default(AxisBase), OnYPropertyChanged));
+
         // Define the Alternative YAxis -- this can be done in XAML
-        public static readonly DependencyProperty AlternativeYAxisProperty = DependencyProperty.RegisterAttached(
-            "AlternativeYAxis", typeof(AxisBase), typeof(SwitchAxisTypeBehavior), new PropertyMetadata(default(AxisBase), OnYPropertyChanged));
+        public static readonly DependencyProperty AlternativeYAxisProperty = DependencyProperty.RegisterAttached
+            ("AlternativeYAxis", typeof(AxisBase), typeof(SwitchAxisTypeBehavior), new PropertyMetadata(default(AxisBase), OnYPropertyChanged));
 
         // Define a boolean attached property which switches between default and alternative YAxis
-        public static readonly DependencyProperty UseAlternateYAxisProperty = DependencyProperty.RegisterAttached(
-            "UseAlternateYAxis", typeof(bool), typeof(SwitchAxisTypeBehavior), new PropertyMetadata(default(bool), OnYPropertyChanged));
+        public static readonly DependencyProperty UseAlternateYAxisProperty = DependencyProperty.RegisterAttached
+            ("UseAlternateYAxis", typeof(bool), typeof(SwitchAxisTypeBehavior), new PropertyMetadata(default(bool), OnYPropertyChanged));
 
         public static void SetDefaultXAxis(DependencyObject element, AxisBase value)
         {
@@ -55,7 +56,7 @@ namespace SciChart.Examples.Examples.ModifyAxisBehaviour
 
         public static AxisBase GetDefaultXAxis(DependencyObject element)
         {
-            return (AxisBase)element.GetValue(DefaultXAxisProperty);
+            return (AxisBase) element.GetValue(DefaultXAxisProperty);
         }
 
         public static void SetAlternativeXAxis(DependencyObject element, AxisBase value)
@@ -65,7 +66,7 @@ namespace SciChart.Examples.Examples.ModifyAxisBehaviour
 
         public static AxisBase GetAlternativeXAxis(DependencyObject element)
         {
-            return (AxisBase)element.GetValue(AlternativeXAxisProperty);
+            return (AxisBase) element.GetValue(AlternativeXAxisProperty);
         }
 
         public static void SetUseAlternateXAxis(DependencyObject element, bool value)
@@ -75,7 +76,7 @@ namespace SciChart.Examples.Examples.ModifyAxisBehaviour
 
         public static bool GetUseAlternateXAxis(DependencyObject element)
         {
-            return (bool)element.GetValue(UseAlternateXAxisProperty);
+            return (bool) element.GetValue(UseAlternateXAxisProperty);
         }
 
         public static void SetDefaultYAxis(DependencyObject element, AxisBase value)
@@ -85,7 +86,7 @@ namespace SciChart.Examples.Examples.ModifyAxisBehaviour
 
         public static AxisBase GetDefaultYAxis(DependencyObject element)
         {
-            return (AxisBase)element.GetValue(DefaultYAxisProperty);
+            return (AxisBase) element.GetValue(DefaultYAxisProperty);
         }
 
         public static void SetAlternativeYAxis(DependencyObject element, AxisBase value)
@@ -95,7 +96,7 @@ namespace SciChart.Examples.Examples.ModifyAxisBehaviour
 
         public static AxisBase GetAlternativeYAxis(DependencyObject element)
         {
-            return (AxisBase)element.GetValue(AlternativeYAxisProperty);
+            return (AxisBase) element.GetValue(AlternativeYAxisProperty);
         }
 
         public static void SetUseAlternateYAxis(DependencyObject element, bool value)
@@ -105,27 +106,29 @@ namespace SciChart.Examples.Examples.ModifyAxisBehaviour
 
         public static bool GetUseAlternateYAxis(DependencyObject element)
         {
-            return (bool)element.GetValue(UseAlternateYAxisProperty);
+            return (bool) element.GetValue(UseAlternateYAxisProperty);
         }
 
         private static void OnXPropertyChanged(DependencyObject d, DependencyPropertyChangedEventArgs e)
         {
             // This is where we switch the axis type depending on boolean
-            var scs = d as SciChartSurface;
-            if (scs == null) return;
-
-            bool useAlternateXAxis = GetUseAlternateXAxis(scs);
-            scs.XAxis = useAlternateXAxis ? GetAlternativeXAxis(scs) : GetDefaultXAxis(scs);
+            if (d is SciChartSurface scs)
+            {
+                bool useAlternateXAxis = GetUseAlternateXAxis(scs);
+                scs.XAxis = useAlternateXAxis ? GetAlternativeXAxis(scs) : GetDefaultXAxis(scs);
+                scs.RenderableSeries.ForEachDo(s => s.XAxisId = scs.XAxis.Id);
+            }
         }
 
         private static void OnYPropertyChanged(DependencyObject d, DependencyPropertyChangedEventArgs e)
         {
             // This is where we switch the axis type depending on boolean
-            var scs = d as SciChartSurface;
-            if (scs == null) return;
-
-            bool useAlternateYAxis = GetUseAlternateYAxis(scs);
-            scs.YAxis = useAlternateYAxis ? GetAlternativeYAxis(scs) : GetDefaultYAxis(scs);
+            if (d is SciChartSurface scs)
+            {
+                bool useAlternateYAxis = GetUseAlternateYAxis(scs);
+                scs.YAxis = useAlternateYAxis ? GetAlternativeYAxis(scs) : GetDefaultYAxis(scs);
+                scs.RenderableSeries.ForEachDo(s => s.XAxisId = scs.XAxis.Id);
+            }
         }
     }
 }
